Create Guess-The-Number.ipynb
The program will first randomly generate a
number unknown to the user. The user needs to guess what that number is. If the user’s guess is wrong, the program should
return some sort of indication as to how wrong (e.g. The number is too high or too low). If the
user guesses correctly, a positive indication should appear.